authorizations

+++ src/services/notes/controller/note-controller.js
@@ -1,13 +1,16 @@
 import NoteRepositories from '../repositories/note-repositories.js';
 import { InvariantError, NotFoundError } from '../../../exception/index.js';
 import response from '../../../utils/response.js';
+import AuthorizationError from '../../../exception/authorization-error.js';
 
 export const createNote = async (req, res, next) => {
-  const { title, tags, body } = req.validated;
+  const { title, body, tags } = req.validated;
+  const { id: owner } = req.user;
   const note = await NoteRepositories.createNote({
     title,
     body,
-    tags
+    tags,
+    owner
   });
   if (!note) {
     return next(new InvariantError('Catatan gagal ditambahkan'));
@@ -17,46 +20,59 @@ export const createNote = async (req, res, next) => {
 };
 
 export const getAllNotes = async (req, res) => {
-  const notes = await NoteRepositories.getNotes();
-  return response(res, 200, 'success', notes);
+  const { id: owner } = req.user;
+  const notes = await NoteRepositories.getNotes(owner);
+  return response(res, 200, 'Catatan sukses ditampilkan', notes);
 };
 
 export const getNoteById = async (req, res, next) => {
   const { id } = req.params;
-  const note = await NoteRepositories.getNotebyId(id);
-
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
+  const note = await NoteRepositories.getNoteById(id);
   if (!note) {
     return next(new NotFoundError('Catatan tidak ditemukan'));
   }
-
   return response(res, 200, 'Catatan sukses ditampilkan', note);
 };
 
-export const editNote = async (req, res, next) => {
+export const editNoteById = async (req, res, next) => {
   const { id } = req.params;
-  const { title, tags, body } = req.validated;
-
+  const {
+    title,
+    body,
+    tags
+  } = req.validated;
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
   const note = await NoteRepositories.editNote({
     id,
     title,
     body,
     tags
   });
-
   if (!note) {
     return next(new NotFoundError('Catatan tidak ditemukan'));
   }
-
   return response(res, 200, 'Catatan berhasil diperbarui', note);
 };
 
-export const deleteNote = async (req, res, next) => {
+export const deleteNoteById = async (req, res, next) => {
   const { id } = req.params;
-  const deleteNote = await NoteRepositories.deleteNote(id);
-
-  if (!deleteNote) {
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
+  const deletedNote = await NoteRepositories.deleteNote(id);
+  if (!deletedNote) {
     return next(new NotFoundError('Catatan tidak ditemukan'));
   }
-
-  return response(res, 200, 'Catatan berhasil dihapus', deleteNote);
+  return response(res, 200, 'Catatan berhasil dihapus', deletedNote);
 };
